+++ Sources/KeyboardKit/Appearance/KeyboardActionCalloutStyle.swift
@@ -90,7 +90,7 @@ public struct KeyboardActionCalloutStyle {
 public extension KeyboardActionCalloutStyle {
     
     /**
-     This standard style aims to mimic the native iOS style.
+     This standard style mimics the native iOS style.
 
      This can be set to change the standard value everywhere.
      */