+++ Sources/KeyboardKit/Appearance/KeyboardButtonBorderStyle.swift
@@ -51,7 +51,7 @@ public extension KeyboardButtonBorderStyle {
     static var noBorder = KeyboardButtonBorderStyle()
     
     /**
-     This standard style aims to mimic the native iOS style.
+     This standard style mimics the native iOS style.
 
      This can be set to change the standard value everywhere.
      */