+++ Sources/KeyboardKit/Appearance/KeyboardButtonShadowStyle.swift
@@ -53,7 +53,7 @@ public extension KeyboardButtonShadowStyle {
     }
     
     /**
-     This standard style aims to mimic the native iOS style.
+     This standard style mimics the native iOS style.
 
      This can be set to change the standard value everywhere.
      */