+++ Sources/KeyboardKit/Appearance/KeyboardButtonStyle.swift
@@ -11,10 +11,13 @@ import SwiftUI
 /**
  This style defines the style of a system keyboard button.
  
- Unlike other styles, this has no `standard` style, since it
+ This type has no `standard` style, since the standard style
  depends on so many factors, like button type, layout config
  etc. Instead, use a ``KeyboardAppearance``to resolve styles
  for certain actions.
+
+ To avoid breaking changes, the style's corner radius is not
+ optional, although it should be.
  */
 public struct KeyboardButtonStyle {