+++ Sources/KeyboardKit/Appearance/KeyboardCalloutStyle.swift
@@ -117,7 +117,7 @@ public struct KeyboardCalloutStyle: Equatable {
 public extension KeyboardCalloutStyle {
     
     /**
-     This standard style aims to mimic the native iOS style.
+     This standard style mimics the native iOS style.
 
      This can be set to change the standard value everywhere.
      */