+++ Sources/KeyboardKit/Autocomplete/AutocompleteToolbarItemBackgroundStyle.swift
@@ -46,8 +46,8 @@ public struct AutocompleteToolbarItemBackgroundStyle {
 public extension AutocompleteToolbarItemBackgroundStyle {
     
     /**
-     This standard style aims to mimic the native iOS style.
-     
+     This standard style mimics the native iOS style.
+
      This can be set to change the standard value everywhere.
      */
     static var standard = AutocompleteToolbarItemBackgroundStyle()