+++ Sources/KeyboardKit/Autocomplete/AutocompleteToolbarSeparatorStyle.swift
@@ -54,7 +54,7 @@ public struct AutocompleteToolbarSeparatorStyle {
 public extension AutocompleteToolbarSeparatorStyle {
     
     /**
-     This standard style aims to mimic the native iOS style.
+     This standard style mimics the native iOS style.
 
      This can be set to change the standard value everywhere.
      */