+++ Sources/KeyboardKit/Autocomplete/AutocompleteToolbarStyle.swift
@@ -61,7 +61,7 @@ public struct AutocompleteToolbarStyle {
 public extension AutocompleteToolbarStyle {
     
     /**
-     This standard style aims to mimic the native iOS style.
+     This standard style mimics the native iOS style.
 
      This can be set to change the standard value everywhere.
      */